requirements.txt
Django==1.7
Pillow==2.5.1
boto==2.36.0
dj-database-url==0.2.2
dj-static==0.0.6
django-appconf==1.0.1
django-compressor==1.4
django-realtime==1.1
django-storages==1.1.8
django-tastypie==0.12.1
django-toolbelt==0.0.1
gunicorn==18.0
httplib2==0.8
mimeparse==0.1.3
oauth2==1.5.211
psycopg2==2.6
python-dateutil==2.1
python-mimeparse==0.1.4
python-openid==2.2.5
six==1.9.0
static3==0.5.1
wsgiref==0.1.2